[Create] 성장 일기에도 바텀 카드 적용
